backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yfinance==0.2.36
pandas==2.1.4
python-dotenv==1.0.0
